perf(charts): lazy-load recharts via next/dynamic
recharts is ~200KB minified, the dominant cost of every chart-using
page chunk. Before this change every detail/overview/stats route
forced the full library into First Load JS, even though the chart is
visible only after the page renders.

Wrap each chart component (PieChart, SimpleLineChart, SimpleBarChart)
in a thin client-only wrapper that defers the actual recharts import
via next/dynamic (ssr: false). The component file's existing default
export and Props type are preserved so no consumer code needs to
change. The loading shell reserves the same height/aspect ratio so
there's no layout shift when the chunk arrives.

First Load JS deltas (production build):

  /overview               314 KB -> 211 KB  (-103 KB)
  /stats                  293 KB -> 189 KB  (-104 KB)
  /top/channels/details   289 KB -> 191 KB  ( -98 KB)
  /top/dms/details        299 KB -> 201 KB  ( -98 KB)
  /top/guilds/details     305 KB -> 207 KB  ( -98 KB)

Charts now arrive in their own chunk, fetched on demand alongside
the page that renders them.
